--- src/components/MediaPlayer/Cards/types.ts
@@ -1,6 +1,0 @@
-import type { MediaPlayer } from "../../../types/MediaPlayer";
-
-export type MediaPlayerProps = {
-    player: MediaPlayer;
-    icon: string;
-};
