--- src/components/MediaPlayer/Table/types.ts
@@ -1,6 +1,0 @@
-import type { MediaPlayer } from "../../../types/MediaPlayer";
-
-export type RowProps = {
-    player: MediaPlayer;
-    icon: string;
-};
