+++ src/types/MediaPlayerDisplay.ts
@@ -1,0 +1,17 @@
+import type { SizeProp } from "@fortawesome/fontawesome-svg-core";
+import type { MediaPlayer, Source } from "./MediaPlayer";
+
+export type MediaPlayersListProps = {
+    players: MediaPlayer[];
+    icons: string[];
+};
+
+export type MediaPlayerProps = {
+    player: MediaPlayer;
+    icon: string;
+};
+
+export type SourceProps = {
+    source: Source;
+    size?: SizeProp;
+};
